refactor: suggestions from code review
